requirements.txt
Django==2.1.1
django-environ==0.4.5
djangorestframework==3.8.2
gunicorn==19.9.0
mysqlclient==1.3.13
pytz==2018.5
